requirements.txt
gssapi
ldap3
certipy-ad